move syncOpenLayers to handler-utils
