+++ src/web/context/handlers/rally-point-handlers.ts
@@ -5,7 +5,7 @@ import { handleMapModeChange } from "../../openlayers/maps/map";
 import { MapModes } from "../../types/openlayers-types";
 import { JaiaContextType, JaiaAction, ButtonNames } from "../../types/context-types";
 import { UNASSIGNED_ID } from "../../utils/constants";
-import { syncOpenLayers } from "../JaiaContext";
+import { syncOpenLayers } from "./handler-utils";
 
 /**
  * Makes call to update the rally point layer
